Move param sub classes to top-level
