+++ src/main/java/uk/co/ramp/covid/simulation/parameters/BuildingDistribution.java
@@ -1,0 +1,35 @@
+package uk.co.ramp.covid.simulation.parameters;
+
+// Defines the number of types of building per N people
+public class BuildingDistribution {
+    public Integer hospitals = null;
+    public PlaceSizeDistribution hospitalSizes = null;
+
+    public Integer schools = null;
+    public PlaceSizeDistribution schoolSizes = null;
+
+    public Integer shops = null;
+    public PlaceSizeDistribution shopSizes = null;
+
+    public Integer offices = null;
+    public PlaceSizeDistribution officeSizes = null;
+
+    public Integer constructionSites = null;
+    public PlaceSizeDistribution constructionSiteSizes = null;
+
+    public Integer nurseries = null;
+    public PlaceSizeDistribution nurserySizes = null;
+
+    public Integer restaurants = null;
+    public PlaceSizeDistribution restaurantSizes = null;
+
+    public boolean isValid() {
+        return hospitalSizes.isValid("hospital")
+                && schoolSizes.isValid("school")
+                && shopSizes.isValid("shop")
+                && officeSizes.isValid("office")
+                && constructionSiteSizes.isValid("construction site")
+                && nurserySizes.isValid("nurseries")
+                && restaurantSizes.isValid("restaurant");
+    }
+}
