+++ src/main/java/uk/co/ramp/covid/simulation/parameters/InfantProperties.java
@@ -1,0 +1,10 @@
+package uk.co.ramp.covid.simulation.parameters;
+
+public class InfantProperties {
+    public Double pAttendsNursery = null;
+
+    public boolean isValid() {
+        //return PopulationParameters.isValidProbability(pAttendsNursery, "pAttendsNursery");
+        return true;
+    }
+}
